+++ src/platform_impl/linux/wayland/touch.rs
@@ -1,17 +1,23 @@
 use std::sync::{Arc, Mutex};
 
+use crate::dpi::PhysicalPosition;
 use crate::event::{TouchPhase, WindowEvent};
 
-use super::{event_loop::EventsSink, window::WindowStore, DeviceId, WindowId};
+use super::{event_loop::EventsSink, make_wid, window::WindowStore, DeviceId};
+
+use smithay_client_toolkit::surface;
 
 use smithay_client_toolkit::reexports::client::protocol::{
     wl_seat,
+    wl_surface::WlSurface,
     wl_touch::{Event as TouchEvent, WlTouch},
 };
 
+// location is in logical coordinates.
 struct TouchPoint {
-    wid: WindowId,
-    location: (f64, f64),
+    surface: WlSurface,
+    x: f64,
+    y: f64,
     id: i32,
 }
 
@@ -31,75 +37,90 @@ pub(crate) fn implement_touch(
                     } => {
                         let wid = store.find_wid(&surface);
                         if let Some(wid) = wid {
+                            let scale_factor = surface::get_dpi_factor(&surface) as f64;
+                            let location =
+                                PhysicalPosition::new(scale_factor * x, scale_factor * y);
+
                             sink.send_window_event(
                                 WindowEvent::Touch(crate::event::Touch {
                                     device_id: crate::event::DeviceId(
                                         crate::platform_impl::DeviceId::Wayland(DeviceId),
                                     ),
                                     phase: TouchPhase::Started,
-                                    location: (x, y).into(),
+                                    location,
                                     force: None, // TODO
                                     id: id as u64,
                                 }),
                                 wid,
                             );
-                            pending_ids.push(TouchPoint {
-                                wid,
-                                location: (x, y),
-                                id,
-                            });
+                            pending_ids.push(TouchPoint { surface, x, y, id });
                         }
                     }
                     TouchEvent::Up { id, .. } => {
                         let idx = pending_ids.iter().position(|p| p.id == id);
                         if let Some(idx) = idx {
                             let pt = pending_ids.remove(idx);
+
+                            let scale_factor = surface::get_dpi_factor(&pt.surface) as f64;
+                            let location =
+                                PhysicalPosition::new(pt.x * scale_factor, pt.y * scale_factor);
+
                             sink.send_window_event(
                                 WindowEvent::Touch(crate::event::Touch {
                                     device_id: crate::event::DeviceId(
                                         crate::platform_impl::DeviceId::Wayland(DeviceId),
                                     ),
                                     phase: TouchPhase::Ended,
-                                    location: pt.location.into(),
+                                    location,
                                     force: None, // TODO
                                     id: id as u64,
                                 }),
-                                pt.wid,
+                                make_wid(&pt.surface),
                             );
                         }
                     }
                     TouchEvent::Motion { id, x, y, .. } => {
                         let pt = pending_ids.iter_mut().find(|p| p.id == id);
                         if let Some(pt) = pt {
-                            pt.location = (x, y);
+                            pt.x = x;
+                            pt.y = y;
+
+                            let scale_factor = surface::get_dpi_factor(&pt.surface) as f64;
+                            let location =
+                                PhysicalPosition::new(x * scale_factor, y * scale_factor);
+
                             sink.send_window_event(
                                 WindowEvent::Touch(crate::event::Touch {
                                     device_id: crate::event::DeviceId(
                                         crate::platform_impl::DeviceId::Wayland(DeviceId),
                                     ),
                                     phase: TouchPhase::Moved,
-                                    location: (x, y).into(),
+                                    location,
                                     force: None, // TODO
                                     id: id as u64,
                                 }),
-                                pt.wid,
+                                make_wid(&pt.surface),
                             );
                         }
                     }
                     TouchEvent::Frame => (),
                     TouchEvent::Cancel => {
                         for pt in pending_ids.drain(..) {
+                            let scale_factor = surface::get_dpi_factor(&pt.surface) as f64;
+                            let location =
+                                PhysicalPosition::new(pt.x * scale_factor, pt.y * scale_factor);
+
                             sink.send_window_event(
                                 WindowEvent::Touch(crate::event::Touch {
                                     device_id: crate::event::DeviceId(
                                         crate::platform_impl::DeviceId::Wayland(DeviceId),
                                     ),
                                     phase: TouchPhase::Cancelled,
-                                    location: pt.location.into(),
+                                    location,
                                     force: None, // TODO
                                     id: pt.id as u64,
                                 }),
-                                pt.wid,
+                                make_wid(&pt.surface),
                             );
                         }
                     }