+++ CHANGELOG.md
@@ -16,6 +16,7 @@
 - On Wayland, fix coordinates in mouse events when scale factor isn't 1
 - On Web, add the ability to provide a custom canvas
 - **Breaking:** On Wayland, the `WaylandTheme` struct has been replaced with a `Theme` trait, allowing for extra configuration
+- On Wayland, fix coordinates in touch events when scale factor isn't 1
 
 # 0.20.0 (2020-01-05)